sort button now works
